Refactor how info is retrieved from the database
Changed the database so that there is only one value for the region, and changed the licensed game being shown to a binary value.

Added a class to pass back the sql statement value needed to get the right combination of licensed games and the region the user has selected.