Search functionality implemented
The search functionality, uses the specific type search, function to pass back games matching a certain genre, sub genre etc. now gets the results based on a search the user asks, I forgot that sql needs to use like, to get results that contain part of the word, as it takes the search term as a literal, so mario returns nothing but like %mario% returns all mario games with mario in the title.