Fixed backstack issue
Can now go from the stats page and show all the genre games.
It needed to pop the backstack to a specific entry in the backstack, now I don't need to repeatedly click the back button to leave the stats page.

Small fixes for the alpha index to show on the owned games fragment list.